+++ src/main/java/com/qcloud/Module/Athena.java
@@ -1,0 +1,15 @@
+package com.qcloud.Module;
+
+/*
+ * Tencent Cloud Financial Intelligent Customer Service.
+ * <p>
+ * see API document: <a
+ * href="https://cloud.tencent.com/document/product/671/14381">
+ * https://cloud.tencent.com/document/product/671/14381</a>
+ * </p>
+ */
+public class Athena extends Base {
+    public Athena(){
+        serverHost = "athena.api.qcloud.com";
+    }
+}